Исправил расчет зарплаты

+++ Agronom/CoreData/CDField + Extensions.swift
@@ -12,7 +12,7 @@ extension CDField {
     func fill(item: Field) {
         self.id = item.id.int16
         self.name = item.name
-        self.size = item.size.int16
+        self.size = item.size.int64
         self.planted = item.planted
     }
